requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.0
pymupdf>=1.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
Pillow>=10.0.0
